% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare4MFA.R
\name{checkFactorialStructure}
\alias{checkFactorialStructure}
\title{Comprueba la estructura factorial de los grupos creados.}
\usage{
checkFactorialStructure(mylistOfGroups)
}
\arguments{
\item{mylistOfGroups}{La lista de grupos de variables a analizar.}
}
\value{
Un marco de datos que muestra la estructura factorial de los grupos.
}
\description{
Comprueba la estructura factorial de los grupos creados.
}
