% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare4MFA.R
\name{extractGroup}
\alias{extractGroup}
\title{Extracts a subgroup of columns from a data matrix}
\usage{
extractGroup(x, pos, vecOfSizes)
}
\arguments{
\item{x}{a data matrix}

\item{pos}{the position of the subgroup to be extracted}

\item{vecOfSizes}{a vector of sizes that defines the number of columns in each continuous subgroup in the data.}
}
\value{
a data frame containing the subgroup of columns
}
\description{
This function extracts a subgroup of columns from a data matrix given the position of the subgroup to be extracted and a vector of sizes that defines the number of columns in each continuous subgroup in the data.
}
