% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare4MFA.R
\name{extractVarNames}
\alias{extractVarNames}
\title{Extract the subset of variables associated with a given label.}
\usage{
extractVarNames(varsAndGroupsDF, label)
}
\arguments{
\item{varsAndGroupsDF}{A data frame with the variables and their associated labels.}

\item{label}{The label for which we want to extract the variables.}
}
\value{
A character vector with the names of the variables associated with the given label.
}
\description{
This function returns the variable names associated with a given label in a data frame.
}
\examples{
varsAndGroupsDF <- varsTable(iris)
varsAndGroupsDF$Group <- c(rep("Sepal",2),rep("Petal",2), "Species")
extractVarNames(varsAndGroupsDF, "Sepal")

}
