% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare4MFA.R
\name{showGroupsinList}
\alias{showGroupsinList}
\title{Muestra información sobre los grupos en una lista de variables.}
\usage{
showGroupsinList(alistOfVars)
}
\arguments{
\item{alistOfVars}{La lista de variables a mostrar.}
}
\description{
Muestra información sobre los grupos en una lista de variables.
}
