% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare4MFA.R
\name{showText}
\alias{showText}
\title{Muestra un texto en la consola con un formato especial.}
\usage{
showText(aText)
}
\arguments{
\item{aText}{El texto que se mostrará en la consola.}
}
\description{
Muestra un texto en la consola con un formato especial.
}
