% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare4MFA.R
\name{varsTable}
\alias{varsTable}
\title{Build a table with information about variables in a data frame}
\usage{
varsTable(x)
}
\arguments{
\item{x}{A data frame containing the variables to be analyzed.}
}
\value{
A data frame with the index, name, class and type of each variable in the data frame.
}
\description{
May be useful when, given a dataset, one wishes to create a table
that will be later used to define subgroups

The function returns a data frame with
- the index,
- name,
- class and
- type of each variable in a data frame.
The type of a variable is determined based on whether it is a factor/character or not.
}
\examples{
varsTable(iris)

}
